codesetslib: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@39220 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/codesetslib/developer/examples/UTF8ToStrHook.c
@@ -54,7 +54,7 @@ struct CodesetsIFace* ICodesets = NULL;
 
 HOOKPROTONH(destFunc, ULONG, struct convertMsg* msg, STRPTR buf)
 {
-    printf("[%3d] [%s]\n",msg->len,buf);
+    printf("[%3d] [%s]\n",(int)msg->len,buf);
 
     if(msg->state == CSV_End)
       printf("\n");
